requirements.txt
GitPython
asyncio
aiohttp